% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{boxplots}
\alias{boxplots}
\title{Single Boxplot}
\usage{
boxplots(df, variable)
}
\arguments{
\item{df}{dataframe provided by the user}

\item{variable}{variable name}
}
\value{
A boxplot of the data
}
\description{
Given a dataset and a variable, creates a boxplot to see if
  there are outliers in the data
}
\examples{
\dontrun{
data(mtcars)
boxplots(mtcars, mtcars$mpg)
}

}
