% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{cat_relation}
\alias{cat_relation}
\title{Plot Categorical Variable Distribution}
\usage{
cat_relation(df, var, y)
}
\arguments{
\item{df}{the dataset from the user}
}
\value{
A ggplot object displaying the distribution of the categorical variable.
}
\description{
Plot Categorical Variable Distribution
}
\examples{
\dontrun{
}
}
