% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{cat_summary_stat}
\alias{cat_summary_stat}
\title{Generate explanatory data analysis for categorical variable}
\usage{
cat_summary_stat(df, var)
}
\description{
Given a numeric variable, this function generate EDA including summary statistics and visualization.
}
