% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_yz.R
\name{categorical_ss}
\alias{categorical_ss}
\title{generate explanatory data analysis for numeric variable}
\usage{
categorical_ss(var)
}
\description{
Given a numeric variable, this function generate EDA including summary statistics and visualization.
}
