% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution.factor}
\alias{distribution.factor}
\title{distribution plot for factor variables}
\usage{
\method{distribution}{factor}(var, df, ...)
}
\arguments{
\item{var}{a factor variable argument.}

\item{df}{a dataframe containing the input variable.}

\item{...}{optional argument for user to specify binwidth (default is 30)}
}
\value{
a histogram of counts for each of the levels for the cyl variable.
}
\description{
Given a factor variable, generates a histogram displaying distribution counts
}
\examples{
data("mtcars")
distribution(mtcars$cyl, mtcars)
}
