% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution.numeric}
\alias{distribution.numeric}
\title{distribution plot for numeric variables}
\usage{
\method{distribution}{numeric}(var, df, ...)
}
\arguments{
\item{var}{a numeric or character argument.}

\item{df}{a dataframe containing the input variable.}

\item{...}{optional argument for user to specify binwidth (default is 30)}
}
\value{
a histogram for the drat variable.
}
\description{
Given a numeric variable, generates a histogram displaying distribution with the option of specifying the number of bins
}
\examples{
data("mtcars")
distribution(mtcars$drat, mtcars)
}
