% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{histograms}
\alias{histograms}
\title{Single Histogram}
\usage{
histograms(df, variable, bins)
}
\arguments{
\item{df}{dataframe provided by the user}

\item{variable}{variable name}

\item{bins}{number of bins the user wants to create}
}
\value{
A histogram distribution of the data provided
}
\description{
Given a dataset and a variable, creates a histogram to see distribution of data
}
\examples{
\dontrun{
data(mtcars)
histograms(mtcars, mtcars$mpg, 5)
}

}
