% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_yz.R
\name{numeric_statistic}
\alias{numeric_statistic}
\title{generate explanatory data analysis for numeric variable}
\usage{
numeric_statistic(df, var)
}
\description{
Given a numeric variable, this function generate EDA including distribution and visualization.
}
