% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{numeric_stats}
\alias{numeric_stats}
\title{EDA for Numerics}
\usage{
numeric_stats(df, var)
}
\arguments{
\item{df}{A dataframe provided by the user}

\item{var}{A variable to compute statistics for}

\item{...}{currently unused}
}
\value{
A list of summary stats
}
\description{
Given a numeric variable, generates basic exploratory data analysis
}
\examples{
data("mtcars")
numeric_stats(mtcars, mtcars$mpg)
}
