% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nataliework.R, R/numeric.R
\name{safe_stats}
\alias{safe_stats}
\title{Safe Summary Stats}
\usage{
safe_stats(x, ...)

safe_stats(x, ...)
}
\arguments{
\item{x}{variable name}

\item{...}{currently unused}

\item{df}{dataframe provided by the user}
}
\value{
A list of summary statistics or an error message stating there are NA values
}
\description{
Given a dataset, computes the summary statistics for the data if
there are no NA values

Given a dataset, computes the summary statistics for the data if
there are no NA values
}
\examples{
data(penguins)
safe_stats(penguins$flipper_length_mm)

}
