% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats.factor}
\alias{summary_stats.factor}
\title{summary statistics for factor variables}
\usage{
\method{summary_stats}{factor}(var, df, na.rm = FALSE)
}
\arguments{
\item{var}{a numeric or character argument.}

\item{df}{a dataframe containing the input variable.}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
a dataframe of summary stats for the mpg variable.
}
\description{
Given a dataset and a variable, the function generates summary statistics based on its type
}
\examples{
\dontrun{
data("iris")
summary_stats(iris$Species, iris)
}
}
