% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats.numeric}
\alias{summary_stats.numeric}
\title{summary statistics for numeric variables}
\usage{
\method{summary_stats}{numeric}(var, df, na.rm = FALSE)
}
\arguments{
\item{var}{a numeric or character argument.}

\item{df}{a dataframe containing the input variable.}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
a dataframe of summary stats for the mpg variable.
}
\description{
Given a dataset and a variable, the function generates summary statistics based on its type
}
\examples{
\dontrun{
data("mtcars")
summary_stats(mtcars$mpg, mtcars)
}
}
