% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\title{distribution plots for variables}
\usage{
distribution(var, df, ...)
}
\arguments{
\item{var}{a numeric or character argument.}

\item{df}{a dataframe containing the input variable.}
}
\value{
a histogram for the mpg variable.
}
\description{
Given a variable, generates plot displaying distribution based on its type
}
\examples{
\dontrun{
data("mtcars")
distribution(mtcars$mpg, mtcars)
}
}
